more fixes

+++ Code/Framework/AzCore/AzCore/Math/Guid.h
@@ -22,7 +22,7 @@ typedef struct _GUID {
 
     _GUID() = default;
 
-    uint32_t       Data1;
+    unsigned long  Data1;
     unsigned short Data2;
     unsigned short Data3;
     unsigned char  Data4[ 8 ];
